fix: bugs in images and big data transfer

+++ crates/egui-states-core/src/image.rs
@@ -12,8 +12,8 @@ pub enum ImageType {
 
 #[derive(Serialize, Deserialize)]
 pub struct ImageInfo {
-    pub image_size: [usize; 2],   // [y, x]
-    pub rect: Option<[usize; 4]>, // [y, x, h, w]
+    pub image_size: [u32; 2],   // [y, x]
+    pub rect: Option<[u32; 4]>, // [y, x, h, w]
     pub image_type: ImageType,
     pub update: bool,
 }
